Cigar.decode broken on Windows
Cigar.decode was calling the C parasail library decode function which mallocs memory.  C parasail does not provide a method for freeing that memory.  ctypes.cdll.msvcrt.free() gave access errors, likely because the parasail.dll was linked to a different runtime.  If the platform is Windows, we use a pure-python implementation.